moved SignedHeader struct from `rpc::endpoint` to `block::signed_header`
